Flatten handlers package
