--- handlers/about_handler/about.go
@@ -1,1 +1,0 @@
-package abouthandler
