+++ routes/router.go
@@ -1,11 +1,7 @@
 package routes
 
 import (
-	adminhandler "anonchihaya.co.uk/handlers/admin_handler"
-	experiencehandler "anonchihaya.co.uk/handlers/experience_handler"
-	homehandler "anonchihaya.co.uk/handlers/home_handler"
-	projecthandler "anonchihaya.co.uk/handlers/project_handler"
-	statichandler "anonchihaya.co.uk/handlers/static_handler"
+	"anonchihaya.co.uk/handlers"
 	"anonchihaya.co.uk/middlewares"
 	"anonchihaya.co.uk/repositories"
 	"github.com/gin-gonic/gin"
@@ -30,10 +26,10 @@ func InitRoutes(r *gin.Engine,
 	admin := r.Group(prefix + "/admin")
 	{
 		admin.POST("", func(ctx *gin.Context) {
-			adminhandler.PostAdminCheck(ctx, domain, admin_pass, key)
+			handlers.PostAdminCheck(ctx, domain, admin_pass, key)
 		})
 		admin.GET("/status", func(ctx *gin.Context) {
-			adminhandler.GetStatusCheck(ctx, key)
+			handlers.GetStatusCheck(ctx, key)
 		})
 	}
 
@@ -42,7 +38,7 @@ func InitRoutes(r *gin.Engine,
 	static.Use(middlewares.KeyChecker(key))
 	{
 		static.POST("/upload-image", func(ctx *gin.Context) {
-			statichandler.UploadImage(ctx, img_path, img_url_prefix)
+			handlers.UploadImage(ctx, img_path, img_url_prefix)
 		})
 	}
 
@@ -50,52 +46,27 @@ func InitRoutes(r *gin.Engine,
 	home := r.Group(prefix + "/home")
 	home.Use(middlewares.KeyChecker(key))
 	{
-		home.GET("", homehandler.GetHomeMsg)
-		home.POST("/upload-profile-img", func(ctx *gin.Context) {
-			homehandler.UploadProfileImg(ctx, img_path, img_url_prefix)
-		})
-		home.GET("/profile-info", func(ctx *gin.Context) {
-			homehandler.GetProfileInfo(ctx, profile_repo)
-		})
-		home.DELETE("/profile-info", func(ctx *gin.Context) {
-			homehandler.DeleteProfileInfo(ctx, profile_repo)
-		})
-		home.POST("/profile-info", func(ctx *gin.Context) {
-			homehandler.PostProfileInfo(ctx, profile_repo)
-		})
-		home.PUT("/profile-info", func(ctx *gin.Context) {
-			homehandler.PutProfileInfo(ctx, profile_repo)
-		})
+		home.GET("", handlers.GetHomeMsg)
+	}
 
-		home.GET("/experience", func(ctx *gin.Context) {
-			homehandler.GetExperiencesShort(ctx, experiences_repo)
+	// * Profile
+	profile := r.Group(prefix + "/profile")
+	profile.Use(middlewares.KeyChecker(key))
+	{
+		profile.POST("/upload-image", func(ctx *gin.Context) {
+			handlers.UploadProfileImg(ctx, img_path, img_url_prefix)
 		})
-
-		home.PUT("/experience/order", func(ctx *gin.Context) {
-			homehandler.PutExperienceOrder(ctx, experiences_repo)
+		profile.GET("", func(ctx *gin.Context) {
+			handlers.GetProfileInfo(ctx, profile_repo)
 		})
-
-		home.POST("/upload-education-img", func(ctx *gin.Context) {
-			homehandler.UploadEducationImg(ctx, img_path, img_url_prefix)
+		profile.DELETE("", func(ctx *gin.Context) {
+			handlers.DeleteProfileInfo(ctx, profile_repo)
 		})
-		home.GET("/education", func(ctx *gin.Context) {
-			homehandler.GetEducations(ctx, educations_repo)
+		profile.POST("", func(ctx *gin.Context) {
+			handlers.PostProfileInfo(ctx, profile_repo)
 		})
-		home.DELETE("/education/:id", func(ctx *gin.Context) {
-			homehandler.DeleteEducation(ctx, educations_repo)
-		})
-		home.POST("/education", func(ctx *gin.Context) {
-			homehandler.PostEducation(ctx, educations_repo)
-		})
-		home.POST("/education/image", func(ctx *gin.Context) {
-			homehandler.PostEducationImg(ctx, educations_repo)
-		})
-		home.PUT("/education", func(ctx *gin.Context) {
-			homehandler.PutEducation(ctx, educations_repo)
-		})
-
-		home.GET("/post/latest", func(ctx *gin.Context) {
-			homehandler.GetPostLatest(ctx, posts_repo)
+		profile.PUT("", func(ctx *gin.Context) {
+			handlers.PutProfileInfo(ctx, profile_repo)
 		})
 	}
 
@@ -104,22 +75,28 @@ func InitRoutes(r *gin.Engine,
 	exp.Use(middlewares.KeyChecker(key))
 	{
 		exp.POST("/upload-experience-img", func(ctx *gin.Context) {
-			experiencehandler.UploadExperienceImg(ctx, img_path, img_url_prefix)
+			handlers.UploadExperienceImg(ctx, img_path, img_url_prefix)
 		})
 		exp.GET("", func(ctx *gin.Context) {
-			experiencehandler.GetAllExperiences(ctx, experiences_repo)
+			handlers.GetAllExperiences(ctx, experiences_repo)
+		})
+		exp.GET("/short", func(ctx *gin.Context) {
+			handlers.GetExperiencesShort(ctx, experiences_repo)
 		})
 		exp.GET("/:id", func(ctx *gin.Context) {
-			experiencehandler.GetExperienceByID(ctx, experiences_repo)
+			handlers.GetExperienceByID(ctx, experiences_repo)
+		})
+		exp.PUT("/order", func(ctx *gin.Context) {
+			handlers.PutExperienceOrder(ctx, experiences_repo)
 		})
 		exp.POST("", func(ctx *gin.Context) {
-			experiencehandler.PostExperience(ctx, experiences_repo)
+			handlers.PostExperience(ctx, experiences_repo)
 		})
 		exp.PUT("", func(ctx *gin.Context) {
-			experiencehandler.PutExperience(ctx, experiences_repo)
+			handlers.PutExperience(ctx, experiences_repo)
 		})
 		exp.DELETE("/:id", func(ctx *gin.Context) {
-			experiencehandler.DeleteExperience(ctx, experiences_repo)
+			handlers.DeleteExperience(ctx, experiences_repo)
 		})
 	}
 
@@ -128,34 +105,67 @@ func InitRoutes(r *gin.Engine,
 	proj.Use(middlewares.KeyChecker(key))
 	{
 		proj.GET("", func(ctx *gin.Context) {
-			projecthandler.GetProjects(ctx, projects_repo)
+			handlers.GetProjects(ctx, projects_repo)
 		})
 		proj.POST("", func(ctx *gin.Context) {
-			projecthandler.PostProject(ctx, projects_repo)
+			handlers.PostProject(ctx, projects_repo)
 		})
 		proj.POST("/update-image-url", func(ctx *gin.Context) {
-			projecthandler.PostProjectImg(ctx, projects_repo)
+			handlers.PostProjectImg(ctx, projects_repo)
 		})
 		proj.PUT("", func(ctx *gin.Context) {
-			projecthandler.PutProject(ctx, projects_repo)
+			handlers.PutProject(ctx, projects_repo)
 		})
 		proj.DELETE("/:id", func(ctx *gin.Context) {
-			projecthandler.DeleteProject(ctx, projects_repo)
+			handlers.DeleteProject(ctx, projects_repo)
 		})
-		proj.GET("/:id/posts", func(ctx *gin.Context) {
-			projecthandler.GetPostsShort(ctx, posts_repo)
+	}
+
+	// * Education
+	education := r.Group(prefix + "/education")
+	education.Use(middlewares.KeyChecker(key))
+	{
+		education.POST("/upload-image", func(ctx *gin.Context) {
+			handlers.UploadEducationImg(ctx, img_path, img_url_prefix)
 		})
-		proj.GET("/post/:id", func(ctx *gin.Context) {
-			projecthandler.GetPost(ctx, posts_repo)
+		education.GET("", func(ctx *gin.Context) {
+			handlers.GetEducations(ctx, educations_repo)
 		})
-		proj.POST("/post", func(ctx *gin.Context) {
-			projecthandler.PostPost(ctx, posts_repo)
+		education.DELETE("/:id", func(ctx *gin.Context) {
+			handlers.DeleteEducation(ctx, educations_repo)
 		})
-		proj.PUT("/post", func(ctx *gin.Context) {
-			projecthandler.PutPost(ctx, posts_repo)
+		education.POST("", func(ctx *gin.Context) {
+			handlers.PostEducation(ctx, educations_repo)
 		})
-		proj.DELETE("/post/:id", func(ctx *gin.Context) {
-			projecthandler.DeletePost(ctx, posts_repo)
+		education.POST("/image", func(ctx *gin.Context) {
+			handlers.PostEducationImg(ctx, educations_repo)
+		})
+		education.PUT("", func(ctx *gin.Context) {
+			handlers.PutEducation(ctx, educations_repo)
+		})
+	}
+
+	// * Post
+	post := r.Group(prefix + "/post")
+	post.Use(middlewares.KeyChecker(key))
+	{
+		post.GET("/latest", func(ctx *gin.Context) {
+			handlers.GetPostLatest(ctx, posts_repo)
+		})
+		post.GET("/project/:id", func(ctx *gin.Context) {
+			handlers.GetPostsShort(ctx, posts_repo)
+		})
+		post.GET("/:id", func(ctx *gin.Context) {
+			handlers.GetPost(ctx, posts_repo)
+		})
+		post.POST("", func(ctx *gin.Context) {
+			handlers.PostPost(ctx, posts_repo)
+		})
+		post.PUT("", func(ctx *gin.Context) {
+			handlers.PutPost(ctx, posts_repo)
+		})
+		post.DELETE("/:id", func(ctx *gin.Context) {
+			handlers.DeletePost(ctx, posts_repo)
 		})
 	}
